Refactor parseFormAndQuery method to return generic map
Updated the parseFormAndQuery method in helper.go file of backend API handlers to return a generic map instead of a map with string keys and values. This change allows for more flexibility in handling different types of form data and query parameters. Also removed the mapToActivities method as it is no longer needed because of this refactor.